src/requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.4.7
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
autotrain-advanced==0.7.126
bitsandbytes==0.43.1
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
codecarbon==2.3.5
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
datasets==2.19.1
debugpy==1.8.2
decorator==5.1.1
deepspeed==0.14.3
diffusers==0.27.2
dill==0.3.8
dnspython==2.6.1
docstring_parser==0.16
einops==0.7.0
email_validator==2.1.2
evaluate==0.4.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
hf_transfer==0.1.6
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
inflate64==1.0.0
invisible-watermark==0.2.0
ipadic==1.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiwer==3.0.3
joblib==1.4.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
loguru==0.7.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
onnx==1.16.1
onnxconverter-common==1.14.0
onnxmltools==1.12.0
onnxoptimizer==0.3.13
onnxruntime==1.15.1
onnxruntime-tools==1.7.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optuna==3.6.1
orjson==3.10.5
packaging==24.0
pandas==2.2.2
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pretty-errors==1.2.25
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
py7zr==0.21.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybcj==1.0.2
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyngrok==7.1.6
pynvml==11.5.0
pyparsing==3.1.2
pyppmd==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
pyzstd==0.16.0
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.31.0
responses==0.18.0
rich==13.7.1
rouge-score==0.1.2
sacremoses==0.1.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.6.18
tiktoken==0.6.0
timm==1.0.3
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1+cu121
torchmetrics==1.4.0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
triton==2.3.1
trl==0.9.4
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
xgboost==2.0.3
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
